build: 0.1.50-PRE-ALPHA
feat: Correctly implement Salts and Hashing
feat: Add Client States [Offline, Online, Away, DND]
feat: Remove name length restriction
feat: Do not load Vue DevTools until after everything else has loaded

+++ packages/server/src/main/kotlin/rebase/FileController.kt
@@ -38,7 +38,7 @@ class FileController() {
         val path = "./storage/user/$user/avatars/avatar_$imageID.webp"
         val file = File(path)
         val existingFiles = File("./storage/user/$user/avatars/").listFiles()
-        if (existingFiles.isNotEmpty()) {
+        if (existingFiles != null && existingFiles.isNotEmpty()) {
             for (existingFile in existingFiles) {
                 existingFile.delete()
             }
